Extended Vehicle with 'Motorised Vehicle' class (e.g. car 'is a' motorised vehicle) which in turn 'has a' engine and fuel tank. Also added vehicle and fuel tank classes.
Also added 'manager' classes for each of the above.

I had problems getting the application to build as I made the mistake of trying to solve two things at the same time and I have no idea which one is the problem. So now I have the classes there, but no initilisation. I will do that one at a time until I get it back.